requirements.txt
# Used by pip to install required python packages
# usage: pip install -r requirements.txt

Flask>=1.0.3
Flask-API>=1.1
Flask-RESTful>=0.3.7
numpy>=1.16.4
pandas>=0.24.2
pip>=9.0.1
python-dateutil>=2.8.0
pytz>=2019.1
redis>=3.2.1
setuptools>=39.0.1
pickle>=4.0
pandas>=0.24.2